happy ball live

+++ app/src/main/java/com/whzl/mengbi/chat/room/message/messageJson/PrizePoolFullJson.java
@@ -21,6 +21,6 @@ public class PrizePoolFullJson {
 
         public long countDownSecond;
         public String periodNumber;
-        public int prizePoolNumber;
+        public long prizePoolNumber;
     }
 }
